Tests

+++ backend/application/room/get_room_query/daos/get_room_dao.go
@@ -8,9 +8,9 @@ type GetRoomDao struct {
 	PlayingSongTitle         *string    `db:"playing_song_title"`
 	PlayingSongAuthor        *string    `db:"playing_song_author"`
 	PlayingSongStartedAtUtc  *time.Time `db:"playing_song_started_at_utc"`
-	PlayingSongLengthSeconds *uint8     `db:"playing_song_length_seconds"`
+	PlayingSongLengthSeconds *uint16    `db:"playing_song_length_seconds"`
 	UserRole                 string     `db:"role"`
 	BoostUsedAtUtc           *time.Time `db:"boost_used_at_utc"`
-	BoostCooldownSeconds     *uint8     `db:"boost_cooldown_seconds"`
+	BoostCooldownSeconds     *uint16    `db:"boost_cooldown_seconds"`
 	SongDaos                 []GetRoomSongDao
 }
