+++ backend/infrastructure/persistance/daos/room_dao.go
@@ -11,7 +11,7 @@ type RoomDao struct {
 	Name                 string    `db:"name"`
 	Password             []byte    `db:"password"`
 	QrCodeHash           []byte    `db:"qr_code_hash"`
-	BoostCooldownSeconds *uint8    `db:"boost_cooldown_seconds"`
+	BoostCooldownSeconds *uint16   `db:"boost_cooldown_seconds"`
 	CreatedAtUtc         time.Time `db:"created_at_utc"`
 	LifespanSeconds      uint32    `db:"lifespan_seconds"`
 }
